Move schemaFor back to tapir package
