+++ playground/src/main/scala/tapir/tests/TestDerivation.scala
@@ -2,7 +2,7 @@ package tapir.tests
 
 import tapir.Schema.SString
 import tapir.Schema
-import tapir.generic.SchemaFor
+import tapir.SchemaFor
 import tapir._
 
 object TestDerivation extends App {
